move generate amendment cli cmd to txs
